fix types

+++ pumps_v1/pumps_v1_digispark/pumps_v1_digispark.ino
@@ -21,7 +21,7 @@ class Pump {
     }
 
   public:
-    Pump(int runTime, int sleepTime, unsigned int pumpPin) {
+    Pump(unsigned long runTime, unsigned long sleepTime, unsigned int pumpPin) {
       timeToSwitchSec = 0;
       runPeriodSec = runTime;
       sleepPeriodSec = sleepTime;
@@ -73,8 +73,8 @@ unsigned long lastMSec = 0;
 void setup() {
   //clock_prescale_set(clock_div_16); // уменьшим частоту
 
-  pumps[0] = new Pump(120, 12 * 3600 - 120, PB0);
-  pumps[1] = new Pump(300, 12 * 3600 - 300, PB1);
+  pumps[0] = new Pump(120, 12UL * 3600UL - 120UL, PB0);
+  pumps[1] = new Pump(300, 12UL * 3600UL - 300UL, PB1);
   for (int i = 0; i < pumpsCount; i++) {
     pumps[i]->initPin();
   }
